Add files via upload
This is the first worksheet concerned with Data Science Track in Udacity at Shell. In this scope of the course, I review and recovered my Pandas, Numpy, Matplotlib and Seaborn knowledges. In this worksheet, I worked with a dataset which aims to analyses No-Show of appointments in particular addresses. I stated 3 base questions and 1 follow up question. I plotted straightforward my visuals relevant to answers. Finally, I am happy to announce and share my work here.